requirements.txt
flask==2.3.0
flask-sqlalchemy==3.1.1
flask-login==0.6.2
werkzeug==2.3.0
mysql-connector-python==8.3.0
bcrypt==4.1.2
python-dotenv==1.0.1
